+++ GgafDx9LibStg/src/jp/ggaf/dx9libstg/util/spline/SplineManufacture.cpp
@@ -9,19 +9,19 @@ SplineManufacture::SplineManufacture(const char* prm_source_file) : GgafObject()
     _pSplSrcCon = (SplineSourceConnection*)((P_GOD)->_pSplSrcManager->connect(prm_source_file));
     _pSplSrc = _pSplSrcCon->use();
     _sp = _pSplSrc->_pSp;
-    _rate_X = 1.0f;
-    _rate_Y = 1.0f;
-    _rate_Z = 1.0f;
+    _rate_X = 1.0;
+    _rate_Y = 1.0;
+    _rate_Z = 1.0;
 }
 SplineManufacture::SplineManufacture(SplineSource* prm_pSplSrc) {
     _pSplSrcCon = NULL;
     _pSplSrc = prm_pSplSrc;
     _sp = _pSplSrc->_pSp;
-    _rate_X = 1.0f;
-    _rate_Y = 1.0f;
-    _rate_Z = 1.0f;
+    _rate_X = 1.0;
+    _rate_Y = 1.0;
+    _rate_Z = 1.0;
 }
-void SplineManufacture::adjustAxisRate(float prm_rate_X, float prm_rate_Y, float prm_rate_Z) {
+void SplineManufacture::adjustAxisRate(double prm_rate_X, double prm_rate_Y, double prm_rate_Z) {
     _rate_X = prm_rate_X;
     _rate_Y = prm_rate_Y;
     _rate_Z = prm_rate_Z;
